+++ src/zepben/evolve/__init__.py
@@ -118,3 +118,5 @@ from zepben.evolve.streaming.grpc.connect import *
 from zepben.evolve.streaming.put.producer import *
 
 from zepben.evolve.util import *
+
+from zepben.evolve.services.network.network_extensions import *